+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - rename the `db_uri` argument of `init_sqlalchemy_unchained` to `database_uri`
 - remove the `ModelMetaOptionsFactory._model_repr` property
 - remove the `validates` decorator
+- add `SessionManager` and `ModelManager` classes
 
 ## v0.5.1 (2018/10/20)